+++ data/src/main/java/app/peter/mos/data/repositories/GoogleRepository.kt
@@ -1,4 +1,0 @@
-package app.peter.mos.data.repositories
-
-class GoogleRepository {
-}