+++ data/src/main/java/app/peter/mos/data/tool/db/DataBase.kt
@@ -1,4 +1,0 @@
-package app.peter.mos.data.tool.db
-
-class DataBase {
-}