+++ domain/src/main/java/app/peter/mos/domain/repository/GoogleRepository.kt
@@ -1,0 +1,6 @@
+package app.peter.mos.domain.repository
+
+interface GoogleRepository {
+    // TODO: Define Google API specific methods
+    suspend fun test(): Result<Unit>
+}